Move classes into main package
