+++ src/main/java/genie/command/AddCommand.java
@@ -1,8 +1,8 @@
 package genie.command;
 
-import genie.Storage;
-import genie.TaskList;
-import genie.Ui;
+import genie.main.Storage;
+import genie.main.TaskList;
+import genie.main.Ui;
 import genie.task.Task;
 
 public class AddCommand extends Command {
