+++ src/main/java/genie/main/Genie.java
@@ -1,4 +1,4 @@
-package genie;
+package genie.main;
 
 import genie.command.Command;
 import genie.exception.DukeException;
