+++ src/main/java/genie/main/MainWindow.java
@@ -1,4 +1,4 @@
-package genie;
+package genie.main;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
